rename variables

+++ BrickCmd/TiledSharp/TmxObjectGroup.cs
@@ -47,8 +47,10 @@ namespace TiledSharp
             {
                 Id = (int?)xObject.Attribute("id") ?? 0;
                 Name = (string)xObject.Attribute("name") ?? String.Empty;
-                X = (int)xObject.Attribute("x");
-                Y = (int)xObject.Attribute("y");
+
+                // you gotta do this?
+                X = (int)(double)xObject.Attribute("x");
+                Y = (int)(double)xObject.Attribute("y");
                 Width = (int?)xObject.Attribute("width") ?? 0;
                 Height = (int?)xObject.Attribute("height") ?? 0;
                 if (xObject.Attribute("class") != null)
